Refactor: Separate bytecode disassembler logic
- I moved the `disassemble_code` function from `main.py` to a new dedicated file `disassembler.py`.
- I updated `index.html` to fetch and load `disassembler.py` in Pyodide instead of using an inline Python script.
- I modified `main.py` to import `disassemble_code` from the new `disassembler.py` to maintain CLI functionality.

This change decouples the core disassembling logic from the web interface and command-line tool, improving modularity and maintainability.
