+++ main.py
@@ -5,26 +5,7 @@ import io
 from contextlib import redirect_stdout
 import argparse
 import importlib.util
-
-def disassemble_code(code_str):
-    """
-    Compile and disassemble a Python code string.
-    
-    Args:
-        code_str (str): Valid Python code as a string
-        
-    Returns:
-        str: The disassembled bytecode
-    """
-    # Compile the code string
-    compiled_code = compile(code_str, '<string>', 'exec')
-    
-    # Capture the output of dis.dis() in a string
-    buffer = io.StringIO()
-    with redirect_stdout(buffer):
-        dis.dis(compiled_code)
-    
-    return buffer.getvalue()
+from disassembler import disassemble_code
 
 def disassemble_function(module_name, function_name):
     """
